added basic server side validation

+++ app.py
@@ -23,23 +23,26 @@ def home():
 @app.post("/add")
 def add():
     title = request.form.get("title")
-    new_todo = Todo(title=title, complete=False)
-    db.session.add(new_todo)
-    db.session.commit()
+    if title:
+        new_todo = Todo(title=title, complete=False)
+        db.session.add(new_todo)
+        db.session.commit()
     return redirect(url_for("home"))
 
 @app.get("/update/<int:todo_id>")
 def update(todo_id):
     todo = db.session.query(Todo).filter(Todo.id == todo_id).first()
-    todo.complete = not todo.complete
-    db.session.commit()
+    if todo:
+        todo.complete = not todo.complete
+        db.session.commit()
     return redirect(url_for("home"))
 
 @app.get("/delete/<int:todo_id>")
 def delete(todo_id):
     todo = db.session.query(Todo).filter(Todo.id == todo_id).first()
-    db.session.delete(todo)
-    db.session.commit()
+    if todo:
+        db.session.delete(todo)
+        db.session.commit()
     return redirect(url_for("home"))
 
 if __name__ == "__main__":
